+++ semcore/date-picker/src/components/PickerAbstract.js
@@ -22,19 +22,23 @@ const i18n = { de, en, es, fr, it, ja, ru, zh, pt, ko, vi };
 
 const INTERACTION_TAGS = ['INPUT'];
 
+const defaultDisplayedPeriod = new Date(new Date().setHours(0, 0, 0, 0));
+
 class PickerAbstract extends Component {
   static displayName = 'DatePicker';
   static style = style;
-  static defaultProps = {
-    i18n,
-    locale: 'en',
-    defaultDisplayedPeriod: new Date(new Date().setHours(0, 0, 0, 0)),
-    defaultValue: null,
-    defaultVisible: false,
-    defaultHighlighted: [],
-    disabled: [],
-    size: 'm',
-  };
+  static defaultProps({ value, defaultValue }) {
+    return {
+      i18n,
+      locale: 'en',
+      defaultDisplayedPeriod: value || defaultValue || defaultDisplayedPeriod,
+      defaultValue: null,
+      defaultVisible: false,
+      defaultHighlighted: [],
+      disabled: [],
+      size: 'm',
+    };
+  }
   static enhance = [i18nEnhance()];
 
   static add = (date, amount, unit) => {
@@ -59,11 +63,9 @@ class PickerAbstract extends Component {
       visible: [
         null,
         (visible) => {
-          if (visible) {
-            const { value } = this.asProps;
-            this.handlers.displayedPeriod(value ? dayjs(value).toDate() : new Date());
-          } else {
+          if (!visible) {
             this.handlers.highlighted([]);
+            this.handlers.displayedPeriod(this.asProps.value || this.props.defaultDisplayedPeriod);
           }
         },
       ],
@@ -192,7 +194,11 @@ class PickerAbstract extends Component {
 
   render() {
     const { styles, Children } = this.asProps;
-    return sstyled(styles)(<Root render={Dropdown}><Children /></Root>);
+    return sstyled(styles)(
+      <Root render={Dropdown}>
+        <Children />
+      </Root>,
+    );
   }
 }