+++ semcore/breadcrumbs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [2.4.1] - 2021-08-24
+
+### Fixed
+
+- Fixed style separator when a custom font-size.
+
 ## [2.4.0] - 2021-04-26
 
 ### Changed