+++ semcore/time-picker/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [2.4.3] - 2021-08-23
+
+### Changed
+
+- Changed height dropdown from `240px` to `180px`.
+
 ## [2.4.2] - 2021-06-04
 
 ### Added